Create 1.1.10
Write a Python program to calculate the average marks for 5 subjects. The program should prompt the user to input the marks for each subject. After receiving the input, it should compute the average marks and then determine the corresponding grade based on the following grading system:



A: 90 - 100
B: 80 - 89
C: 70 - 79
D: 60 - 69
F: Below 60


The program should display the average marks up to 2 decimal places and the assigned grade.